refactor: refactor file structure and component structure for the application
