+++ frontend/src/app/admin/page.tsx
@@ -1,5 +1,12 @@
-import Navbar from "@/component/dashboard/navbar";
+import Navbar from "@/components/navbar/navbar";
 
 export default function adminPage() {
-    return <Navbar />;
+    return (
+        <div>
+            <Navbar />
+            <main>
+                <h1>Admin Page</h1>
+            </main>
+        </div>
+    );
 }
