+++ frontend/src/app/dashboard/page.tsx
@@ -1,5 +1,9 @@
-import Dashboard from "@/component/dashboard/Dashboard";
+import Dashboard from "@/components/dashboard/Dashboard";
 
 export default function page() {
-    return <Dashboard />;
+    return (
+        <div>
+            <Dashboard></Dashboard>
+        </div>
+    );
 }
