+++ frontend/src/app/page.tsx
@@ -1,11 +1,16 @@
-import { SignIn } from "@/component/authentication/SignIn";
-import { SignUp } from "@/component/authentication/SignUp";
+import { SignIn } from "@/components/authentication/SignIn";
+import { SignUp } from "@/components/authentication/SignUp";
+import Navbar from "@/components/navbar/navbar";
 
 export default function Home() {
     return (
-        <>
+        <div>
+            <Navbar />
+            <main>
+                <h1>Content Goes Here</h1>
+            </main>
             <SignUp />
             <SignIn />
-        </>
+        </div>
     );
 }
