--- frontend/src/component/dashboard/dashboardContent.tsx
@@ -1,7 +1,0 @@
-export default function Content() {
-    return (
-        <div>
-            <p>Test Content</p>
-        </div>
-    );
-}
